Created a posts section and made some improvements
A new version of the model with a table of posts was
created.
NOTE: if you already have a database and it contains some
data, then download it to a file and delete the tables,
when trying to migrate from the first version to the second,
due to the relationship of the first table with the new one,
an error may occur.

It became possible to write posts and also view them. A new
decorator has been created to control the access of
unauthorized users to the post creation address.

Also improved some points in the code, such as the
readability of authorization functions, decorators and
a couple of minor changes